Connect to SQL Server database; connection string is in environment variable 'dbconn'.

+++ Models/Reservations.cs
@@ -6,8 +6,8 @@ namespace hotelGetReservation.Models
         public string Id { get; set; }
         public DateTime CheckIn { get; set; }
         public int NumberOfNights { get; set; }
-        public double Rate { get; set; }
-        public double TotalCost {
+        public decimal Rate { get; set; }
+        public decimal TotalCost {
             get {
             return NumberOfNights * Rate;
             }
